Implement new CBOR protocol
This new protocol includes a major refactor of the driver for better
extensibility.

This also includes a new CBOR encoder and decoder, which at this stage
is not entirely complete and contains assertions.

In fact, a lot of the current code contains assertions related to the
protocol, which need to be replaced with something to be caught by the
client handler.